Refactor screenshot helper functions into separate module.
